Move prepare_invoke to Account
